[*]科技位面:属性为零的属性不予显示

+++ src/circe/circe_client/World3Dlg.cpp
@@ -50,26 +50,26 @@ void CWorld3Dlg::_ConvertCardToStirng(const FGCard& card, CString& csConvertStri
         break;
     }
 
-    csConvertString.Format(L"%s  ",cardType);
-	if((BYTE)0!=card.byRevert)
+    csConvertString.Format(L"%s  ", cardType);
+	if (0 != card.byRevert)
 	{
-		csConvertString.AppendFormat(L"»Ø¸´Öµ:%d",card.byRevert);
+	  csConvertString.AppendFormat(L"»Ø¸´Öµ:%d", card.byRevert);
 	}
-	if((BYTE)0!=card.bySuckBlood)
+	if (0 != card.bySuckBlood)
 	{
-		csConvertString.AppendFormat(L" ÎüÑªÖµ:%d",card.bySuckBlood);
+	  csConvertString.AppendFormat(L" ÎüÑªÖµ:%d", card.bySuckBlood);
 	}
-	if((DWORD)0!=card.dwSpecial)
+	if (0 != card.dwSpecial)
 	{
-		csConvertString.AppendFormat(L" ÃâÉË:%d",card.dwSpecial);
+	  csConvertString.AppendFormat(L" ÃâÉË:%d", card.dwSpecial);
 	}
-	if((BYTE)0!=card.byPressAttack)
+	if (0 != card.byPressAttack)
 	{
-		csConvertString.AppendFormat(L" Ç¿Ï®:%d",card.byPressAttack);
+	  csConvertString.AppendFormat(L" Ç¿Ï®:%d", card.byPressAttack);
 	}
-	if((BYTE)0!=card.byCriticalStrike)
+	if (0 != card.byCriticalStrike)
 	{
-		csConvertString.AppendFormat(L" ±©»÷:%d",card.byCriticalStrike);
+	  csConvertString.AppendFormat(L" ±©»÷:%d", card.byCriticalStrike);
 	}
 }
